refactor: move test to new folder
